refactor(parser): change current parameter type from int to size_t for improved type safety

+++ src/parser/Parser.hpp
@@ -36,8 +36,7 @@ public:
 private:
     std::vector<Token>                         tokens;
     std::vector<std::unique_ptr<AtomizerBase>> atomizers;
-
-    int current = 0;
+    size_t                                     current = 0;
 
     bool  isAtEnd() const;
     Token peek() const;
