+++ src/parser/atomizer/atomizers/LoadAtomizer.hpp
@@ -31,7 +31,7 @@ namespace Opal {
 
 class LoadAtomizer : public AtomizerBase {
 public:
-    LoadAtomizer(int& current, std::vector<Token>& tokens);
+    LoadAtomizer(size_t& current, std::vector<Token>& tokens);
     bool                      canHandle(TokenType type) const override;
     std::unique_ptr<NodeBase> atomize() override;
 };
